ICAO-By-ECI/ICAO-By-ECI/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
astroid==2.4.2
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.3
category-encoders==2.6.3
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
customtkinter==5.2.2
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
discord==2.3.1
discord.py==2.3.1
distlib==0.3.7
Django==4.2.4
docutils==0.20.1
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.3.2
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
gb==0.1.4
google==3.0.0
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.22.0
idna==2.10
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
ipywidgets==8.1.2
isoduration==20.11.0
isort==4.3.21
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
jupyterlab_widgets==3.0.10
keyring==24.2.0
kiwisolver==1.4.4
lab==8.2
lazy-object-proxy==1.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplot==0.1.9
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mistune==3.0.1
more-itertools==10.1.0
multidict==6.0.4
mysql-connector-python==8.3.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.25.2
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pbr==5.11.1
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2023.7.23
pkginfo==1.9.6
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.6.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.15.1
pyloco==0.0.139
pyodbc==5.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.2
QtPy==2.4.1
quran==1.0.0
readme-renderer==40.0
referencing==0.30.2
requests==2.32.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
simplejson==3.19.2
SimpleWebSocketServer==0.1.2
six==1.15.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.2
stevedore==5.1.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tk==0.1.0
toml==0.10.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.10.0
twine==4.0.2
txt2tags==3.9
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.25.10
ushlex==0.99.1
virtualenv==20.24.2
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
voila==0.5.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==12.0
Werkzeug==2.3.6
widgetsnbextension==4.0.10
wrapt==1.12.1
yarl==1.9.2
zipp==3.16.2
